finish diplay methods

+++ dotNet5782_4199_5185/BL/DroneAtParcel.cs
@@ -9,7 +9,7 @@ namespace IBL.BO
     public class DroneAtParcel
     {
         public int Id { get; set; }
-        public int Battary { get; set; }
+        public double Battary { get; set; }
         public Location CurrentLocation { get; set; }
 
         public override string ToString()
